+++ config/application.rb
@@ -16,5 +16,8 @@ module CardswiprNew
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.
     config.eager_load_paths += ["#{config.root}/lib"]
+
+    # Set time zone
+    config.time_zone = 'Eastern Time (US & Canada)'
   end
 end