+++ db/migrate/20190803001334_change_check_in_to_checked_in.rb
@@ -1,0 +1,5 @@
+class ChangeCheckInToCheckedIn < ActiveRecord::Migration[5.2]
+  def change
+    rename_column :attendances, :check_in, :checked_in_at
+  end
+end