+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_06_24_160724) do
+ActiveRecord::Schema.define(version: 2019_08_03_001334) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -25,7 +25,7 @@ ActiveRecord::Schema.define(version: 2019_06_24_160724) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.bigint "event_id"
-    t.datetime "check_in"
+    t.datetime "checked_in_at"
     t.index ["event_id"], name: "index_attendances_on_event_id"
   end