+++ spec/models/attendance_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Attendance, type: :model do
     it { is_expected.to validate_presence_of(:email) }
     it { is_expected.to validate_presence_of(:net_id) }
     it { is_expected.to validate_presence_of(:upi) }
-    it { is_expected.to validate_presence_of(:check_in) }
+    it { is_expected.to validate_presence_of(:checked_in_at) }
   end
 
   describe 'Associations' do