+++ spec/system/attendance_spec.rb
@@ -62,7 +62,7 @@ RSpec.describe 'Attendance', type: :system do
     it 'has the right content' do
       data = generate_attendances
       click_on('Export CSV')
-      export = 'first_name,last_name,email,net_id,upi,check_in'
+      export = 'first_name,last_name,email,net_id,upi,checked_in_at'
       export += data.map { |n| export_row_for(n) }.join
       expect(page.html).to have_content(export)
     end
@@ -93,6 +93,7 @@ RSpec.describe 'Attendance', type: :system do
 
   def export_row_for(attendance)
     "\n" + [attendance.first_name, attendance.last_name, attendance.email,
-            attendance.net_id, attendance.upi, attendance.check_in].join(',')
+            attendance.net_id, attendance.upi, attendance.checked_in_at]
+           .join(',')
   end
 end